Changing conditional statements to a switch for Main Activity options menu.
Fixing a bug in SelectionAdapter that has plagued me for a couple of weeks. I don't fully understand it, but changing my use of the 'convertView' parameter in the 'getView()' method to 'v' seems to work.

Changing 'item' Item object to a more readable 'newItem' in the 'updateItemList' method.

Changing the method name 'getItemFromServer' to 'serverSync' in 'ServerDataSource'. This will eventually be the entry point from a stub manual sync process, as well as my eventual attempt at implementing a SyncAdapter.