MainActivity Fixing a bug in the 'onClickListener' for each row. The timestamp was being updated in the database, but not the list of Items in memory. This was causing a logic error in the JSONObjectRequest method where it would not properly send updated Items to the server.
ItemDataSource
Modifying the 'updateItem' method to also take a timestamp of the update as a parameter instead of calculating it itself.
'getCurrentTime' method is now public.

ServerDataSource
Creating an 'itemsToJSONAry' method to create the properly formed JSON object consisting of a timestamp, and all updated items to be sent to the server.