ItemDataSource Creating a 'cursorToList' method to reuse code that iterates over a cursor to build a list of Items.
SelectionAdapter
Removing an if-statement from previous commit that deleted an item from the SelectionAdapter if it was
'isDeleted' while in the 'getView()' method. Horrible idea.

ServerDataSource
'updateItems', the Item array of items that have been updated since our 'lastCheckIn', is now created
via a select statement to the database.
Cleaning up some logic that dictated what to do with items to be updated sent from the server.