ItemContentProvider Update method now provides the current timestamp of the update.
MainActivity
Updating the OnClickListener of 'MainActivity' to appropriately mark and update the ContentProvider
accordingly.
Updating the projection in the 'onCreateLoader' method to query all columns for each Item row.

SelectionAdapter
Updating 'newView' and 'bindView' to check if an Item is 'isMarked', and sets the strikethrough
decoration accordingly.
Also moving the 'getView' method up within the Class to put custom helper methods below
the class overrides.