[ItemLoader] Implementing custom loader to return a list of Item objects rather than the SimpleCursorLoader. Also moving the ItemObserver class out of MainActivity to here.
[ItemUtils]
Public helper class providing cursor-to-item conversion.

[SelectionAdapter]
Reverting this class back to its state in a previous commit. This change returns the adapter to extend an ArrayList instead of a SimpleCursorAdapter so as to be in line with the new Loader changes.

[MainActivity]
Includes necessary changes to enable the activity to work with the updates to the loader, adapter, and Item observer.
Added two methods, setIsMarked() and setMarkedItems(). The former updates a SparseBoolean array mapping arraylist positions to Item.isMarked values. The latter iterates over a list of Items setting values with setIsMarked().

[ServerDataSource]
Moving the cursorToList() and cursorToItem() methods to ItemUtils. Upading these method calls where appropriate.